frontend/requirements.txt
streamlit==1.53.1

requests==2.32.5
python-dotenv==1.2.1

pillow==12.1.0

pandas==2.3.3
numpy==2.4.1

altair==6.0.0
pydeck==0.9.1
narwhals==2.15.0

protobuf==6.33.4
pyarrow==23.0.0

tenacity==9.1.2
toml==0.10.2
tornado==6.5.4
